rename project to pysecsan as pysan is taken in pypi.org
