+++ projects/pysan/pysecsan/LICENSE
@@ -11,12 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from setuptools import setup, find_packages
-
-setup(
-    name='pysan',
-    version='1.0',
-    author='David Korczynski',
-    author_email='david@adalogics.com',
-    packages=find_packages(exclude="tests")
-)
