+++ projects/pysan/pysecsan/pysecsan.egg-info/PKG-INFO
@@ -1,0 +1,18 @@
+Metadata-Version: 2.1
+Name: pysecsan
+Version: 0.2.0
+Summary: Sanitizers to detect security vulnerabilities at runtime.
+Author: David Korczynski
+Author-email: David Korczynski <david@adalogics.com>
+Project-URL: Homepage, https://github.com/google/oss-fuzz
+Project-URL: Bug Tracker, https://github.com/google/oss-fuzz/issues
+Classifier: Programming Language :: Python :: 3
+Classifier: License :: OSI Approved :: Apache Software License
+Classifier: Operating System :: OS Independent
+Requires-Python: >=3.7
+Description-Content-Type: text/markdown
+License-File: LICENSE
+
+# pysecsan
+
+Security sanitizers for vulnerability detection during runtime.
