+++ projects/pysan/pysecsan/pysecsan.egg-info/SOURCES.txt
@@ -1,0 +1,12 @@
+LICENSE
+README.md
+pyproject.toml
+setup.py
+pysecsan/__init__.py
+pysecsan/command_injection.py
+pysecsan/redos.py
+pysecsan/sanlib.py
+pysecsan.egg-info/PKG-INFO
+pysecsan.egg-info/SOURCES.txt
+pysecsan.egg-info/dependency_links.txt
+pysecsan.egg-info/top_level.txt
