+++ projects/pysan/pysecsan/tests/poe/ansible-runner-cve-2021-4041/fuzz_ansible_runner.py
@@ -17,8 +17,8 @@
 import os
 import sys
 import atheris
-import pysan
-pysan.pysan_add_hooks()
+import pysecsan
+pysecsan.pysan_add_hooks()
 
 import pexpect
 import ansible_runner
