+++ projects/pysan/pysecsan/tests/poe/libvcs-cve-2022-21187/fuzz_libvcs.py
@@ -17,8 +17,8 @@
 import os
 import sys
 import atheris
-import pysan
-pysan.pysan_add_hooks()
+import pysecsan
+pysecsan.pysan_add_hooks()
 
 from libvcs.shortcuts import create_repo
