+++ projects/pysan/pysecsan/tests/subprocess_popen_injection.py
@@ -16,34 +16,36 @@
 import os
 import sys
 import atheris
+import subprocess
 
-import pysan
+import pysecsan
 
 
 def list_files_perhaps(param, magicval):
-    if magicval == 1337:
-        try:
-            os.system(param)
-        except ValueError:
-            pass
-    elif magicval == 1338:
-        os.system("exec-san")
-    elif magicval == 1339:
-        os.system("ls -la FROMFUZZ")
-    else:
-        return 2
+    try:
+        subprocess.Popen(" ".join(['ls', '-la', param]), shell=True)
+        os.system(param)
+    except ValueError:
+        pass
 
 
 def TestOneInput(data):
     fdp = atheris.FuzzedDataProvider(data)
-    list_files_perhaps(
-        fdp.ConsumeUnicodeNoSurrogates(24),
-        fdp.ConsumeIntInRange(500, 1500)
-    )
+
+    if fdp.ConsumeIntInRange(1, 10) == 5:
+        list_files_perhaps(
+            "FROMFUZZ",
+            fdp.ConsumeIntInRange(500, 1500)
+        )
+    else:
+        list_files_perhaps(
+            fdp.ConsumeUnicodeNoSurrogates(24),
+            fdp.ConsumeIntInRange(500, 1500)
+        )
 
 
 def main():
-    pysan.pysan_add_hooks()
+    pysecsan.pysan_add_hooks()
 
     atheris.instrument_all()
     atheris.Setup(sys.argv, TestOneInput, enable_python_coverage=True)
@@ -52,4 +54,3 @@ def main():
 
 if __name__ == "__main__":
     main()
-
